Use Worker API to validate plugins
This lets us avoid doing custom classloading in ValidatePlugins.
