+++ subprojects/plugin-development/plugin-development.gradle.kts
@@ -42,6 +42,7 @@ dependencies {
     implementation(project(":plugins"))
     implementation(project(":pluginUse"))
     implementation(project(":publish"))
+    implementation(project(":workers"))
 
     implementation(library("slf4j_api"))
     implementation(library("groovy"))
